Decouple DevFS from specific driver modules
Each driver now owns its VfsNode impl and registers with devfs via a
generic register_device(name, node) API. DevFS no longer imports any
driver modules.
